move shell into its own module
