+++ .travis.yml
@@ -1,8 +1,9 @@
 language: scala
 
 scala:
-  - 2.11.8
   - 2.10.6
+  - 2.11.11
+  - 2.12.2
 
 jdk:
   - oraclejdk8
