+++ modules/shell/src/main/scala/tuco/shell/builtins.scala
