+++ modules/shell/src/main/scala/tuco/shell/command.scala
