+++ modules/shell/src/main/scala/tuco/shell/commands.scala
