+++ modules/shell/src/main/scala/tuco/shell/commandshell.scala
