+++ modules/shell/src/main/scala/tuco/shell/complete.scala
