+++ modules/shell/src/main/scala/tuco/shell/package.scala
@@ -1,0 +1,10 @@
+package tuco
+
+import tuco.Tuco.SessionIO
+
+package object shell {
+
+  def runShell[A](init: Session[A]): SessionIO[Session[A]] =
+    CommandShell.run(init)
+
+}
