+++ modules/shell/src/main/scala/tuco/shell/session.scala
