+++ project/plugins.sbt
@@ -1,7 +1,5 @@
 addSbtPlugin("org.tpolecat"      %  "tut-plugin"            % "0.4.6")
-// addSbtPlugin("com.eed3si9n"      %  "sbt-unidoc"            % "0.3.1")
 addSbtPlugin("com.jsuereth"      %  "sbt-pgp"               % "1.0.0")
 addSbtPlugin("com.github.gseitz" %  "sbt-release"           % "1.0.0")
 addSbtPlugin("org.xerial.sbt"    %  "sbt-sonatype"          % "0.5.0")
-// addSbtPlugin("org.scalastyle"    %% "scalastyle-sbt-plugin" % "0.8.0")
 addSbtPlugin("com.fortysevendeg" %  "sbt-microsites"        % "0.3.1")
